Add files via upload
I understand it is supposed to be a one button game but I used 3 arrows because the code for it is identical and it adds more of a challenge to the game. The object of the game is to make it to the other side without crashing into any police cars or driving off the road. I was trying to have the car images with the transparent background but I wasn't able to find a way to have the background's stay transparent once they were in processing. To go forward you use the right arrow key, to turn left the control is the top arrow and to turn right the control is the down arrow.