Add files via upload
I made a driving game called on the run. The goal is get to the other side while avoiding police and staying on the road. Going off the road or hitting any police cars ends the game. I understand it was supposed to be a one button game but the code for the 3 arrow keys are almost identical. To drive forward you press the right arrow key, turn right by pressing the down arrow key and turn left by pressing the up arrow key. The car images I used have transparent backgrounds but I couldn't get them to have the transparent backgrounds once they were in processing.